feat(font): Non-integer point sizes

+++ include/ghostty.h
@@ -379,7 +379,7 @@ typedef struct {
   ghostty_platform_u platform;
   void* userdata;
   double scale_factor;
-  uint8_t font_size;
+  float font_size;
   const char* working_directory;
   const char* command;
 } ghostty_surface_config_s;
